🌟 stablize feature
