+++ src/main/kotlin/com/theapache64/stackzy/ui/feature/appdetail/AppDetailScreen.kt
@@ -2,12 +2,9 @@ package com.theapache64.stackzy.ui.feature.appdetail
 
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.animation.core.tween
-import androidx.compose.desktop.LocalAppWindow
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Tab
@@ -20,22 +17,13 @@ import androidx.compose.ui.draw.rotate
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.res.imageResource
-import androidx.compose.ui.text.AnnotatedString
-import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.platform.Font
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import com.theapache64.stackzy.data.local.AnalysisReport
-import com.theapache64.stackzy.data.local.GradleInfo
-import com.theapache64.stackzy.data.local.Platform
 import com.theapache64.stackzy.data.remote.Library
-import com.theapache64.stackzy.ui.common.*
-import com.theapache64.stackzy.ui.theme.*
-import com.theapache64.stackzy.ui.util.Preview
+import com.theapache64.stackzy.ui.common.CustomScaffold
+import com.theapache64.stackzy.ui.common.FullScreenError
+import com.theapache64.stackzy.ui.common.LoadingText
 import com.theapache64.stackzy.util.R
 
-private const val GRID_SIZE = 4
 
 @Composable
 fun AppDetailScreen(
@@ -54,8 +42,7 @@ fun AppDetailScreen(
         report!!.appName!!
     }
 
-    // Calculating item width based on screen width
-    val appItemWidth = (LocalAppWindow.current.width - (CONTENT_PADDING_HORIZONTAL * 2)) / GRID_SIZE
+
 
     CustomScaffold(
         title = title,
@@ -66,7 +53,7 @@ fun AppDetailScreen(
         },
         topRightSlot = {
             report?.let {
-                Badge("${it.apkSizeInMb} MB")
+                Badge("APK SIZE: ${it.apkSizeInMb} MB")
             }
         }
     ) {
@@ -106,9 +93,11 @@ fun AppDetailScreen(
                 )
 
                 if (selectedTabIndex == 0) {
-                    LibsUsed(report!!, appItemWidth, onLibrarySelected)
+                    // Libraries Tab
+                    Libraries(report!!, onLibrarySelected)
                 } else {
-                    // show more info
+                    // More Info tab
+                    MoreInfo(report!!)
                 }
 
             }
@@ -116,59 +105,6 @@ fun AppDetailScreen(
     }
 }
 
-@Composable
-private fun LibsUsed(
-    report: AnalysisReport,
-    appItemWidth: Int,
-    onLibrarySelected: (Library) -> Unit
-) {
-    if (report.libraries.isEmpty()) {
-        // No libraries found
-        val platform = report.platform
-        if (platform is Platform.NativeKotlin || platform is Platform.NativeJava) {
-            // native platform with libs
-            FullScreenError(
-                title = "We couldn't find any libraries",
-                message = "But don't worry, we're improving our dictionary strength. Please try later",
-                image = imageResource("drawables/guy.png")
-            )
-        } else {
-            // non native platform with no libs
-            FullScreenError(
-                title = "// TODO : ",
-                message = "${report.platform.name} dependency analysis is yet to support",
-                image = imageResource("drawables/code.png")
-            )
-        }
-    } else {
-
-        LazyColumn {
-            items(
-                items = if (report.libraries.size > GRID_SIZE) {
-                    report.libraries.chunked(GRID_SIZE)
-                } else {
-                    listOf(report.libraries)
-                }
-            ) { appSet ->
-                Row {
-                    appSet.map { app ->
-
-                        // GridItem
-                        Selectable(
-                            modifier = Modifier.width(appItemWidth.dp),
-                            data = app,
-                            onSelected = onLibrarySelected
-                        )
-                    }
-                }
-
-                Spacer(
-                    modifier = Modifier.height(10.dp)
-                )
-            }
-        }
-    }
-}
 
 @Composable
 private fun LoadingAnimation(loadingMessage: String) {
